refactor matrix implementation for sparse matrix implementation

+++ cpp/src/topology/matrix.cpp
@@ -19,46 +19,46 @@ class BinaryMatrix {
 
     // matrix consisting of elements from the binary field {0, 1}
     private:
-        size_t rows; // number of rows 
-        size_t cols; // number of columns
-        size_t col_ints; // how many integers do we need per row?
+        unsigned int rows; // number of rows 
+        unsigned int cols; // number of columns
+        unsigned int col_ints; // how many integers do we need per row?
 
         vector<uint64_t> data; // 1D vector containing 64-bit integers
 
         // these are computed after row reduction
         bool row_reduced = false;
-        size_t rank = 0;
-        size_t nullity = cols;
+        unsigned int rank = 0;
+        unsigned int nullity = cols;
        
-        void add_rows(size_t target, size_t source) {
+        void add_rows(unsigned int target, unsigned int source) {
             // add source row to target row in place
-            size_t target_start = target * col_ints;
-            size_t source_start = source * col_ints;
+            unsigned int target_start = target * col_ints;
+            unsigned int source_start = source * col_ints;
 
             // XOR every pair of integers in the row
-            for (size_t k = 0; k < col_ints; k++) {
+            for (unsigned int k = 0; k < col_ints; k++) {
                 data[target_start + k] ^= data[source_start + k];
             }
         }
 
-        void swap_rows(size_t row1, size_t row2) {
+        void swap_rows(unsigned int row1, unsigned int row2) {
             if (row1 == row2) {return;}
             // swap the values in each row
-            size_t start1 = row1 * col_ints;
-            size_t start2 = row2 * col_ints;
+            unsigned int start1 = row1 * col_ints;
+            unsigned int start2 = row2 * col_ints;
 
-            for (size_t k = 0; k < col_ints; k++) {
+            for (unsigned int k = 0; k < col_ints; k++) {
                 std::swap(data[start1 + k], data[start2 + k]);
             }
         }
 
-        size_t find_pivot(size_t i) const {
+        unsigned int find_pivot(unsigned int i) const {
             // return the position of the pivot at row i
             // fast look up by looking at trailing zeros
             auto start = i*col_ints;
 
             // loop through the integers in the given row
-            for (size_t k = 0; k < col_ints; k++) {
+            for (unsigned int k = 0; k < col_ints; k++) {
                 if (data[start + k] != 0) {
                     // return the number of trailing zeros plus the column that we started looking from
                     return (k*64) + (__countr_zero(data[start + k]));
@@ -73,19 +73,19 @@ class BinaryMatrix {
 
     public:
         // constructor initializes to zero matrix of given dimensions
-        BinaryMatrix(size_t r, size_t c) 
+        BinaryMatrix(unsigned int r, unsigned int c) 
         : rows(r), 
           cols(c), 
           col_ints((c+63)/64),
           data((c+63)/64*r, 0) {}
 
         // return value at (i, j) 
-        bool operator()(size_t i, size_t j) const { 
+        bool operator()(unsigned int i, unsigned int j) const { 
             if (i >= rows || j >= cols) {throw std::out_of_range("i,j out of bounds.");}
 
             // locate the bit corresponding to the i, j entry
-            size_t chunk_i = (i*col_ints) + (j/64);
-            size_t bit_offset = j % 64;
+            unsigned int chunk_i = (i*col_ints) + (j/64);
+            unsigned int bit_offset = j % 64;
 
             // shift the bits, AND it against 1
             bool value = (data[chunk_i] >> bit_offset) & 1ULL;
@@ -101,7 +101,7 @@ class BinaryMatrix {
             // iterate through each chunk and XOR them
             // make a copy of the matrix
             BinaryMatrix result = *this;
-            for (size_t i = 0; i < data.size(); i++) {
+            for (unsigned int i = 0; i < data.size(); i++) {
                 // XOR adds component-wise
                 result.data[i] ^= other.data[i];
             }
@@ -119,17 +119,17 @@ class BinaryMatrix {
             // initialize resulting matrix product
             BinaryMatrix result(rows, other.cols);
         
-            for (size_t i = 0; i < rows; i++) {
-                for (size_t j = 0; j < cols; j++) {
+            for (unsigned int i = 0; i < rows; i++) {
+                for (unsigned int j = 0; j < cols; j++) {
 
                     // If A(i, j) is 1, we XOR the entire j-th row of B into the i-th row of Result
                     if ((*this)(i, j) == 1) {
 
-                        size_t result_start = i * result.col_ints;
-                        size_t b_start = j * other.col_ints;
+                        unsigned int result_start = i * result.col_ints;
+                        unsigned int b_start = j * other.col_ints;
                     
                         // Fast chunk-by-chunk XOR
-                        for (size_t chunk = 0; chunk < result.col_ints; chunk++) {
+                        for (unsigned int chunk = 0; chunk < result.col_ints; chunk++) {
                             result.data[result_start + chunk] ^= other.data[b_start + chunk];
                         }
                     }
@@ -144,7 +144,7 @@ class BinaryMatrix {
             }
 
             // test for equality component-wise
-            for (size_t i = 0; i < data.size(); i++) {
+            for (unsigned int i = 0; i < data.size(); i++) {
                 if (data[i] != other.data[i]) {
                     return false;
                 }
@@ -161,10 +161,10 @@ class BinaryMatrix {
         }
 
         // access number of rows and columns, if needed
-        size_t nrows() const {return rows;}
-        size_t ncols() const {return cols;}
+        unsigned int nrows() const {return rows;}
+        unsigned int ncols() const {return cols;}
 
-        BinaryMatrix row_i (size_t i) const {
+        BinaryMatrix row_i (unsigned int i) const {
             // return row i as a Binary Matrix
             BinaryMatrix output(1, cols);
 
@@ -179,12 +179,12 @@ class BinaryMatrix {
         void add_row(vector<bool>& new_row) {
             // add a new row to the matrix
             if (new_row.size() != cols) {throw std::invalid_argument("New row dimensions does not match.");}
-            for (size_t k = 0; k < col_ints; k++) {
+            for (unsigned int k = 0; k < col_ints; k++) {
                 // add a new row to fill
                 data.push_back(0);
             }
 
-            for (size_t j = 0; j < new_row.size(); j++) {
+            for (unsigned int j = 0; j < new_row.size(); j++) {
                 // update the bits
                 if (new_row[j]) {(*this)(data.size()-1, j);} 
             }
@@ -192,14 +192,14 @@ class BinaryMatrix {
 
         // make the value at (i, j) a one
         // useful for constructing the boundary maps
-        void one(size_t i, size_t j) {
+        void one(unsigned int i, unsigned int j) {
             if (i >= rows || j >= cols) {throw std::out_of_range("i,j out of bounds.");}
 
             // matrix is changed, it isn't row_reduced anymore
             row_reduced = false;
 
-            size_t chunk_i = (i*col_ints) + (j/64);
-            size_t bit_offset = j % 64;
+            unsigned int chunk_i = (i*col_ints) + (j/64);
+            unsigned int bit_offset = j % 64;
 
             // equal to 1 only at the same position as bit_offset
             uint64_t mask = 1ULL << bit_offset;
@@ -209,14 +209,14 @@ class BinaryMatrix {
         }
 
         // make the value at (i, j) a zero
-        void zero(size_t i, size_t j) {
+        void zero(unsigned int i, unsigned int j) {
             if (i >= rows || j >= cols) {throw std::out_of_range("i,j out of bounds.");}
 
             // matrix is changed, it isn't row reduced anymore
             row_reduced = false;
 
-            size_t chunk_i = (i*col_ints) + (j/64);
-            size_t bit_offset = j % 64;
+            unsigned int chunk_i = (i*col_ints) + (j/64);
+            unsigned int bit_offset = j % 64;
 
             // equal to 1 only at the same position as bit_offset
             uint64_t mask = 1ULL << bit_offset;
@@ -226,7 +226,7 @@ class BinaryMatrix {
             data[chunk_i] &= ~mask;
         }
 
-        pair<size_t, size_t> row_reduce() {
+        pair<unsigned int, unsigned int> row_reduce() {
             // row reduce the matrix to RREF, compute and return rank and nullity.
 
 
@@ -240,15 +240,15 @@ class BinaryMatrix {
                 nullity = cols;
             }
 
-            size_t curr_row = 0;
+            unsigned int curr_row = 0;
 
             while (curr_row < rows) {
-                size_t min_pivot = -1;
-                size_t best_row = -1;
+                unsigned int min_pivot = -1;
+                unsigned int best_row = -1;
 
                 // go looking for the best row to use (i.e. leftmost pivot)
-                for (size_t i = curr_row; i < rows; i++) {
-                    size_t pivot = find_pivot(i);
+                for (unsigned int i = curr_row; i < rows; i++) {
+                    unsigned int pivot = find_pivot(i);
                     if (pivot < min_pivot) {
                         min_pivot = pivot;
                         best_row = i;
@@ -266,7 +266,7 @@ class BinaryMatrix {
                 nullity--;
 
                 // zero out all the other rows in this column
-                for (size_t i = 0; i < rows; i++) {
+                for (unsigned int i = 0; i < rows; i++) {
                     if (i != curr_row && (*this)(i, min_pivot) == 1) {
                         add_rows(i, curr_row);
                     }
@@ -285,11 +285,11 @@ class BinaryMatrix {
             // print the matrix to the console
             if (verbose) {
                 cout << "BinaryMatrix([" << endl;
-                for (size_t i = 0; i < rows; i++) {
+                for (unsigned int i = 0; i < rows; i++) {
 
                     cout << "  "; // offset
 
-                    for (size_t j = 0; j < cols; j++) {
+                    for (unsigned int j = 0; j < cols; j++) {
 
                         // print the row
                         cout << (*this)(i, j);
@@ -305,9 +305,9 @@ class BinaryMatrix {
                 cout << "Rows: " << rows << ", Columns: " << cols << "." << endl;
             } else {
                 // not verbose, just print numeric values
-                for (size_t i = 0; i < rows; i++) {
+                for (unsigned int i = 0; i < rows; i++) {
 
-                    for (size_t j = 0; j < cols; j++) {
+                    for (unsigned int j = 0; j < cols; j++) {
 
                         // print the row
                         cout << (*this)(i, j);
